+++ src/MWelcome/Config.h
@@ -14,11 +14,12 @@ enum class WelcomeType
 
 struct Config
 {
-    int         version       = 1;
-    std::string type          = "toast";
-    std::string toast_title   = "§lWelcome to the §bServer§r§l!§r";
-    std::string toast_content = "Enjoy your stay, §o§6{}§r!";
-    std::string msg_content   = "§lWelcome to the §bServer§r, §o§6{}§r!";
+    int         version             = 2;
+    bool        send_to_all_players = false;
+    std::string type                = "toast";
+    std::string toast_title         = "§lWelcome to the §bServer§r§l!§r";
+    std::string toast_content       = "Enjoy your stay, §o§6{}§r!";
+    std::string msg_content         = "§lWelcome to the §bServer§r, §o§6{}§r!";
 
     [[nodiscard]] WelcomeType getType() const
     {